requirements.txt
Pillow
cairosvg
PyGithub
